EntityIdHtmlLinkFormatter: do not check existence of non-local titles
This prevents EntityIdHtmlLinkFormatter to format foreign
entities as deleted. Formatter should not consider whether pages
related to foreign entities exist or not. Title::exists always
returns false for such pages but Formatter should still not be
marking them as deleted. It should link to the entity page
on the foreign repository, and let the third-party to display
an error page if needed.

This also fixes the behaviour for entities for which entity title
lookup would return null as a title. Such entities are considered
deleted.

Bug: T157003
Change-Id: I4cf8e2e78b0bec3e6bf8ce006f1ffd6af73512c4