Update patch set 20
Patch Set 20:

@Aleksey The question is what the intended contract of the ChangeOpDeserializer is, and if we want to focus on testing against the contract, or testing implementation details.

In my mind, the contract of e.g. LabelChangeOpDeserializer is NOT that it creates a ChangeOpLabel. In my mind, the contract is that it creates a ChangeOp that *does* the correct thing! 

In other words: I believe the deserializers should be specified in therms of how they change an entity, not in terms of which ChangeOps they create.

This means that the concrete ChangeOp created is an implementation detail that should NOT be tested. So we cannot rely on the unit tests for the ChangeOp, but have to test the patching process end-to-end.

Patch-set: 20
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0