Parse foreign entity page links in LinkBeginHookHandler
Links to entities from foreign repositories are of the form
[[foo:Special:EntityPage/P123]], where "foo" stands for the
repository "prefix".

Given such links, handler checks whether "foo" is a known
repository, and given that it constructs entity ID
in the form "foo:P123". Label and description of the
foreign entity are used in the link if present.
ID part of the link uses "foo:P123" form.

Note: current implementation assumes that repository prefix
in the foreign entity ID is equivalent to interwiki. This
assumption is currently valid but it might no longer be
accurate in the future.

Bug: T157004
Change-Id: Ic7166f86df31c0143ea00ed8efad3d8d5a75d10d