Update patch set 10
Patch Set 10: Code-Review-1

the elastic search documents contain stuff like:

"labels": {
  "en": [
    "mercury",
    "Hg",
    "quicksilver",
    "hydrargyrum",
    "element 80"
  ],
  "nb": [
    "kvikksølv",
    "Hg"
   ],
  "fr": [
    "mercure",
    "Hg"
  ],

So, things listed as "labels" that really are not labels? they are aliases?

I think it would be better if we could more explicitly separate these in the mappings and documents. Or at least not misleadingly call these "labels"

Patch-set: 10
Label: Code-Review=-1