Introduce RepositoryDefinitions
RepositoryDefinitions is a service providing access to settings
of configured repository (both "local" repository and possible
"foreign" repositories).

Definitions are provided in the form of map from repository
names (strings, with empty-string key standing for the local
repository) to an array of repository-specific settings.

RepositoryDefinitions will be instantiazed by Client and/or
Repo based on the configuration of the Wikibase instance,
and passed to services requiring the knowledge on the
repository settings (e.g. DispatchingServiceFactory,
RepositoryServiceContainerFactory, etc).

Bug: T157570
Change-Id: Id23db330f20d0e3bcd62669607db1b12e70049b6