Update patch set 1
Patch Set 1: Code-Review-2

Note that this patch is not yet ready.
I still need to think which of keys would be obligatory, and which would be allowed to be skipped. Plus I need to figure out the best way to test wfWarn is fired.

Also I think it makes to see this class actually being used. For this I will soon make a patch actually using this with dispatching service infrastructure we have currently.

Patch-set: 1
Label: Code-Review=-2