Update patch set 10
Patch Set 10:

they are both labels and aliases. First element is the label, the rest are aliases. 

> I think it would be better if we could more explicitly separate these

Since the search should look in both, I (and David) don't think it's better to have separate fields, on the contrary, we'll get better results if we have the single field. As I said, it is easy to distinguish between primary label and alias - primary label is always first.

Patch-set: 10