Update patch set 15
Patch Set 15: Code-Review+2

(2 comments)

This looks good and makes sense to me. Obviously, the number of abstraction layers is making this a little hard to gasp, but that seems ok to me.

A minor thing that wont work this way: Searching for an Entity by giving the actual id (with or without repo).

Patch-set: 15
Reviewer: Hoo man <50@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2