Update patch set 3
Patch Set 3: -Code-Review

> How about injecting `LoggerInterface` and checking that `$this->logger->warning(...)` was called. And for production we can create some handler for Monolog delegating/mapping PSR log levels to mediawiki functions.

That sounds like some plan. But as wfWarn also triggers a PHP error, I am checking for this now.
In current state if there are such duplicate settings it means there is something wrong with configuration (which e.g. might lead to unexpected/wrong search results in the system), so we want to warn user about it loud. So what I think we want this class to do in such case is to fail. It makes sense it also prints some notice at the same time but what I am checking now is the error, not the notice.

Thanks for the suggestion nonetheless!

Patch-set: 3
Label: Code-Review=0