Update patch set 4
Patch Set 4:

(1 comment)

Thanks for working on this, but I'm afraid I do have the same problems as before. Without code that actually uses this I can't tell if this is in a mergeable state or not. And the class level and parameter documentation is still not sufficient, in my opinion.

Patch-set: 4
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0