Update patch set 1
Patch Set 1:

> <meta name="twitter:card" content="summary">

It is Twitter specific. The task (and especially patch) is about Open Graph

>  What will happen it the TextExtracts extension starts setting the same meta tags?

It doesn't (as far as I understood). Why should it start? If so, then what to do?
Can't we solve it when the problem arises?


I don't think that patch should be blocked by these issues

Patch-set: 1