Update patch set 15
Patch Set 15: Code-Review+1

(1 comment)

Code looks very good (even if some methods are extremely long, but this is not an issue that is introduced with this patch). I think we can merge this as it is and continue working on this later.

No +2 because I might have missed something and want to give others a chance to have a look.

Patch-set: 15
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1