Update patch set 1
Patch Set 1:

Regarding adding twitter card. I think that needs to be done in PageImages extension not here.
If you check my TextExtract patch, you'll see it needs to be enabled and it's disabled by default. We are going to disable that for Wikidata because we are handling it on our own.
Regarding tests, the tests for ViewEntityAction is super basic so I can't add tests in one or two lines (I tried) or even with one patch. There should be a follow up patch on the tests only and make it cover most of the lines in ViewEntityAction class first and then adds this.

Patch-set: 1