Update patch set 4
Patch Set 4:

The concept of a "standard entity type" does not exist any more. For this test it does not any difference if it uses "kitten" or "item". If you remove the supported entity type from the test setup, the test below will fail. This does not change with my patch.

Patch-set: 4