Update patch set 1
Patch Set 1: Code-Review-2

JSON, SQL, SVG, TXT and XML files? Why?

Sorry, but this is a waste of yours and reviewers time. We do have a PHPCS rule in place that enforces this for PHP files. It does not make much of a difference, but makes working with PHP code in an IDE a little bit smoother. But for these other file types? Why?

Patch-set: 1