Update patch set 1
Patch Set 1:

It is a good style to end every text file with a newline. Missing newlines at the end can cause problems. For example when last line of an SQL file is a comment (--) and it does not have a newline at the end and you concatenate this SQL file with an other SQL file then the first line of the second SQL file becomes also a comment. Missing newlines at the end of TXT files are annoying when you cat this file in a shell because your next prompt is not at the beginning of a line.

Patch-set: 1