Update patch set 1
Patch Set 1:

> It is Twitter specific.

All three tags are Twitter specific. Without the Twitter tag, the two og tags do not serve any purpose and should not be added. Therefor the -1.

> I think that needs to be done in PageImages extension not here.

No. The Twitter card tag has nothing to do with the PageImages extension. PageImages sets the og:image tag, but this exact tag is optional in Twitter cards. The non-optional tags are here in this patch.

Patch-set: 1