Update patch set 2
Patch Set 2: Code-Review-1

This patch is not "clean" if it outputs tags that are known to have no effect. And this is still a problem, even if PS2 added the previously missing tag. The code is still arranged in a way that will output all kinds of non-functional combinations of tags: either a lonely og:description, or an og:title with the Twitter card tag attached, but no og:description.

I can see two possible solutions:
* Output og:title if there is a title. Output og:description if there is a description. But output the Twitter card tag only if both og:title and og:description are there.
* Output all three tags only if both title and description can be filled in.

Both solutions would minimize the "HTML header garbage" we output. There is a problem, though: By doing this we encode very specific knowledge about how Twitter works. The moment Twitter changes it's semantics, we might need to update our code.

Other services are not affected because we are adding the og:title and og:description dupes *exclusively* for Twitter.

Patch-set: 2
Label: Code-Review=-1