Update patch set 2
Patch Set 2:

> It reveals the relation between pieces of code.

Ok, this is an argument I did not considered enough. However, what "relation" would that be?

This is very similar to a string like 'en'. Where would a constant for this language code live? What's the relation?

Currently the constant for the 'item' string lives in the Item class. But a lot of the code that uses the 'item' string does not need to know about the Item class. The "entity type identifier" is just that. There is no strong relation to a single class. 'item' is a concept that does not live in a single file. It would be better to declare this constant via configuration in a flat PHP file, as we do it with the constants for content types.

Let's say we have such an ITEM_TYPE constant. What would you learn when you follow it and find it in a file …/vendor/wikibase/data-model/WikibaseDataModel.php that's otherwise empty? Sure, you learn that there is a relation to the DataModel component. I tend to question the value of this learning, considering the other arguments we shared.

> The reason why the test […] was introduced […] was to test that when a repository instance A is configured to use entities from some other repository B […]

But this is still the case. The entity ID under test is prefixed, referring to a remote repository. My change just means that both repositories use Items.

Patch-set: 2