Update patch set 1
Patch Set 1:

It's not broken, it seems incomplete. The caching logic seems off, that's what the -1 is for. I fear that if we merge this as-is, we'll forget about looking into that issue. But having a ticket for that would be enough. I'll make one so this patch can be merged.

Patch-set: 1