Update patch set 1
Patch Set 1:

(1 comment)

The caching logic that is $this->entitySerializers.

Anyway, I just realized that newDispatchingSerializer doesn't exist in EntitySerializerFactory, that'S why I'm keeping the -1. Also, the deserializer factory methods in WikibaseRepo should also make use of the new features, I think. Can be in a separate patch, though.

Patch-set: 1