Update patch set 2
Patch Set 2:

> Let's say we have such an ITEM_TYPE constant. What would you learn
 > when you follow it and find it in a file …/vendor/wikibase/data-model/WikibaseDataModel.php
 > that's otherwise empty? Sure, you learn that there is a relation to
 > the DataModel component. I tend to question the value of this
 > learning, considering the other arguments we shared.

* It says that it is type. Not random string, not repository prefix that contains items, not whatever else.
For important things I would rather suggest to create a class representing this. It is not just a string, 
for example, concatenation does not make sense for entity type.
* I don't want to remember if it should be written in upper case, lower case or steady caps; use spaces, dashes or underscores (if we speak in general).
* Item as a concept has bunch of classes related to it (Item, ItemId, ItemChange, ItemDeserializer...) and I
would expect that all those will be located somewhere near this type constant. If you put it into plain file
then this file should be in the same directory, if in the class then in the same namespace.
* Also some other minor benefits like: using static analysis tools to understand your real code structure and dependencies

Patch-set: 2