SqlEntityInfoBuilder should ignore entity IDs from other repositories
SqlEntityInfoBuilder should only be used to get information for
entities from repository which database the builder is able to
access.
As database queries fetching entity information, at least partly,
rely on numeric part of EntityIDs, builder accepting
foreign entity IDs could lead to returning wrong results.

Therefore SqlEntityInfoBuilder instance becomes aware of which
repository it is configured to be used with, and only gets EntityInfo
for entities belonging to this repository. Entities from other
repositories are ignored

SqlEntityInfoBuilderFactory also becomes aware of the repository
name, so it instantiates SqlEntityInfoBuilders correctly.

Change-Id: Iba5badca3bd0af0274e6069a2e2f4e7f8080e6bf