Update patch set 1
Patch Set 1: Code-Review-1

Thiemo, you're right, there are two cases to consider here.

We talked briefly with Aleksey about when it can happen there is no service container for the repository and we only focused on the case when there is something wrong with configuration, ie. the repo X is configured in repo defintions but there is somehow no service container for that repo. with current implementation of RepositoryServiceContainer, its wiring file etc it is quite theoretical but in general if getServiceMap is trying to get a container for known repository but the container is not there there is something wrong with the configuration. And it makes sense for the factory to throw an exception. System operating in a state with the invalid configuration is asking for trouble, so it seems reasonable and valid to me that the exception/fatal is emitted when system realizes something wrong's going on.

But, as Thiemo says, this class also implements the EntityStoreWatcher interface and provides methods that are called when changes to entities happen. And it is not impossible (maybe not very likely in case of wikidata.org but pretty possible for third-party installation and I'd say almost certain in case of the development environment) that they're called for the entity coming from the repository that is _currently_ not configured. It could be, for instance, an entity once imported from some repository but then the repo was removed from the configuration. Such entries probably should have been migrated somehow so database is consistent but wikibase generally does not guarantee such thing. So yeah, I believe while it is OK to remove a null check from getServiceMap, watcher methods ie. entityUpdated, entityDeleted, and redirectUpdated should keep this check as we don't want users to have their installation broken if they decide to change their configuration but still have data related to the old config in their database. Such cases should be gracefully handled by Wikibase.

I am to blame for missing this scenario from our consideration, sorry for that!

Patch-set: 1
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1