Update patch set 5
Patch Set 5:

And a general remark related to that: custom entity types might also want to add their "attributes" to the result of wbsearchentities. I believe we'd like to have a general mechanism allowing so. A hook might be one option (not sure if the nicest one, though).
Once we have such a mechanism, this code should/could probably be migrated, so only when there are properties, it would be called.

Patch-set: 5