Update patch set 4
Patch Set 4: Code-Review+1

(1 comment)

Sorry for breaking this! I don't how I've managed to miss this.

The change looks sane to me. Haven't verified manually yet.
I am not sure what would be the best way to deal with problems with this setting (not anything against this patch, I mean more of a general problem of Client depending on Repo's setting and so, see the TODO in WikibaseClient). But this is beyond this patch.

Patch-set: 4
Label: Code-Review=+1