Update patch set 10
Patch Set 10:

still not convinced that combining labels and aliases like this is the best way, though if we want to do this then suggest we call the field something other than "labels" that more accurately represents what the field contains.

the contents and structure of the search index are exposed externally (e.g. cirrusdump), and should be somewhat clearer what it contains

Copying labels and aliases to some "all" field is perhaps reasonable.

Patch-set: 10