Update patch set 1
Patch Set 1:

This patch can be merged as it is without being perfectly "clean". For a start the new option should be applied to the code that actually stores entity blobs to the database. All other places that use serializers can be set to do the same one by one (with an announcement every time). If the numeric format is gone everywhere, we can remove this code.

If this patch is approved, we can move the class to the serialization component, and move the constant to the factory. But I'm not sure if this is worth it if this is going to die anyway.

Patch-set: 1
Label: Verified=0