Call WikibaseEntityNamespaces hook earlier in setup
Call the hook, via a static method, before the WikibaseRepo
factory is constructed. (which optionally includes services
from the WikibaseClient factory which in turn optionally
needs some settings etc. from the repo)

The client factory also needs entity namespaces and
the static method with the hook is the best (or workable)
solution in this case so there is not circular dependency
between the two factories, and the client is aware
of the extra Wikibase content namespaces.

Bug: T157919
Change-Id: Ie6f22cdb6dc54b47fd64ab16b21ded2961d71ec6