Call WikibaseEntityNamespaces hook earlier in setup
Call the hook, via a static method, before the WikibaseRepo
factory is constructed. (which optionally includes services
from the WikibaseClient factory which in turn optionally
needs some settings etc. from the repo)

The client factory also needs entity namespaces and
the static method with the hook is the best (or workable)
solution in this case so there is not circular dependency
between the two factories, and the client is aware
of the extra Wikibase content namespaces.

The patch also removes the fixLegacyContentModelSetting
code, previously used in normalizing the entityNamespaces
setting. That code was temporarily needed for deploying changes
to the content model settings. Wikidata config has been updated
to the new settings and it's been six months for others
to adapt. (imho enough time)

https://github.com/wikimedia/operations-mediawiki-config/commit/81535563b1463d8c154a7dd10970c16706e0de19#diff-01f4cd9b9585049d676035224d9b5464

Bug: T157919
Change-Id: Ie6f22cdb6dc54b47fd64ab16b21ded2961d71ec6