Fix violations caused by to generic, not needed EntityIdComposers
This fixes a series of actual, quite critical code smells. Note that an
EntityIdComposer may return any EntityId it feels like. But all the code
places I'm touching in this patch are strongly bound to a single, specific
entity type:

* What's the point of injecting a generic composer if we already know the
  entity type?
* Some of this code even assumes it always gets an ItemId or PropertyId and
  will fail hard if this is not the case. But the guarantee is gone because
  a generic composer can not give one.

This basically reverts Ibd22680. It appears some of this was done with the
wrong idea that newFromNumber is "bad" and must be destroyed. But there is
nothing wrong with it. We will never remove it from the ItemId and PropertyId
classes. With the introduction of the Int32EntityId interface we even wrote a
specification for this.

The only problem is if newFromNumber is called on an *unknown* entity type.

Bug: T151578
Change-Id: Ic4ee11f98bc73272909ce6b7125301b334e4ca8d