ChangeOpDeserializer throws ChangeOpDeserializerException on invalid input
ChangeOpException is a RuntimeException individual exception
throw when they cannot apply their changes to the given Entity
(e.g. trying to update a statement which does not exist).

Those kind of issues are different from the case when provided
serialized change request is invalid as in it has a wrong
format/structure. In such cases ChangeOpDeserializer should
be rather throwing ChangeOpDeserializerException which is
a InvalidArgumentException. In such cases ChangeOp would
not even be created, thus ChangeOpException related to the
application of the change is not relevant here.

Change-Id: Ifa22ed7ea2defa017144dbf5969f51ec4177525d