[WIP] EntityDataRetrievalServiceFactory used as EntityStoreWatcher?
This is a quite heavy type safety violation. What this line of code means:

* It assumes all classes that implement the EntityDataRetrievalServiceFactory
  interface also implement the EntityStoreWatcher interface.
* Currently there is only one EntityDataRetrievalServiceFactory implementation.
* This means we can as well ditch the interface and type hint against the
  specific DispatchingServiceFactory implementation.

An other solution is to add an if ( instanceof EntityStoreWatcher ) check. But
this means the registerWatcher() call might be skipped, and I can not tell if
this is acceptable (probably not).

The patches that introduced this are Id5dfc9e (T155633) and I15d33dd (T155623).

Change-Id: I1895da624e4f5693f05b158243efe03254783bae