Remove type safety violation from OutputPageBeforeHTMLHookHandler
This code assumes all entities do have labels and descriptions. This patch does
not solve this problem. All it does is moving the existing type safety violation
from OutputPageBeforeHTMLHookHandler one level deeper to
EntityViewPlaceholderExpander.

This is a step in a series of changes, uploaded as a seperate patch to minimize
the diff and make reviewing this as easy as possible.

Change-Id: Ie723c0256018fcc600e0a5b3d46a68da36a72c69