Fix TermView components assuming all entities have labels/descriptions
I found two places in the code that still assumed all entity types implement
the Labels/DescriptionsProvider interfaces. This is the most minimal patch I
could do to fix this.

What this patch basically does is replacing Labels/DescriptionProviders with
TermLists, and AliasesProvider with AliasGroupLists.

I am *not* lifting the limitation that the term views only work if they get
labels and descriptions. It may make sense to make descriptions optional, but
I want to do this in a later patch. With this patch labels and descriptions
fall back to empty lists if an entity type does not support them.

I tried to minify this patch or split it into multiple smaller patches
but failed. If you have a specific suggestion, please tell me or (better)
create a patch and let me rebase mine on top of yours.

Bug: T134735
Change-Id: Icbee448c0fee7fbf66f68456d637fb1aeecee9f8