Remove not needed ChangeListTransformer interface
This was never "used". There was never more than one implementation.
There was never more than one user. All code is in the Client component
anyway, so dependency inversion is not an argument to have this interface.

With I1334bca (see the comment there) it became obvious this interface is
"overspecified". The documentation says it is a generic transformer for
whatever Changes, when in reality it was used for EntityChanges (this
means it was expected to not only accept but also *return* EntityChanges).

Change-Id: Ic5c738e5675ca5721c7c8dc03ea7d383fc7d5b59