Update patch set 1
Patch Set 1: Code-Review+1

Similar code is in WikibaseClient. Should we remove this as well?

I talked to Lydia and we think it's ok to remove this, even if we both think 12 months would be better.

I will send an email to the mailing list announcing this, because it might break third-party installations.

No +2 to give others a little time to see this.

Patch-set: 1