Update patch set 2
Patch Set 2:

The patch seems now technically correct.
I am still not convinced it is change for better. What is wrong with using the class that can build IDs from a string part for all known entity types in those places, while the use of this service
I am not saying that "using newFromNumber is wrong", to be clear.

I wouldn't be surprised if we soon realized that ChangesSubscriptionTableBuilder and/or SqlSiteLinkConflictLookup.php could be also used to acces foreign repo's database thus we would then add ItemId::newForeignFromNumber?

I'd very much like to be convinced. What is said in the commit message does not really convince me.

 > Some of this code even assumes it always gets an ItemId or PropertyId and will fail hard if this is not the case. But the guarantee is gone because a generic composer can not give one.

Should the code relying on the ID assume it itself then? And also if one ends up in the situation that when she asks EntityIdComposer to compose entity ID of type, let's say 'property' but it returns a LexemeId because of completely wrong configuration, it seems to me that it would not be bad if system was failing loud in such case.

Patch-set: 2