Update patch set 2
Patch Set 2:

What is the point of using a generic factory if all the code I'm touching in this patch already knows what type of EntityId it needs? This is like injecting an, let's say, namespace lookup for a single getEntityType( 140 ) with a static number. You should inject the entity type instead. What this does is not only making the code smaller and easier to read, most importantly: when there is a $entityType constructor parameter instead of an $namespaceLookup it's much easier to see that this code is supposed to deal with a single namespace only. Same for the ItemIds and PropertyIds in question here. A PropertyInfoTableBuilder for example will never deal with other entity types than Properties.

Patch-set: 2