Remove all optional SEO meta tags from diff pages
This patch does two things:
1. Add missing tests.
2. Remove all meta tags from diff pages. This even includes the classic
   <meta description> tag. Reasoning: Diff pages do *not* represent the
   content. A description like "German author" does *not* describe the
   contents of a diff page.

I considered changing the "twitter:card" algorithm but decided to keep it.
The current algorithm always duplicates the title, if there is one, and
the description, if there is one. This is *not* useful for Twitter, but may
be useful for other services. It's also more consistent.

Bug: T51859
Change-Id: I660083ae703c21cbc3ba69e695c1e23497f3b67d