Update patch set 3
Patch Set 3:

> it's a good practice to add those tags regardless even if it's a duplicate of other tags because it's for different platforms.

Even if this statement is technically correct, stating that a meta tag is worth something just because of it's nature is pretty much the typical "SEO snakeoil" argument I hear since the late 1990s, and that proved to be bad advice since then. What's still missing in this discussion is the name of a service other than Twitter that is relevant for us and does not understand <title> or <meta description>, but does understand og:title or og:description. Can you name one? That would be very helpful.

Patch-set: 3