Update patch set 1
Patch Set 1:

This looks like it could work. I'd suggest to remove/adjust documentation referencing those global variables in *.entitytypes.php files, and also in docs/entitytypes.wiki in the same patch.

Why is the global used in the end? Is it something from the times before extensions like MediaInfo were using a hook to register their custom entity types? Or did Wikibase let users register entity types by fiddling with the global in their LocalSettings? That would be a point in favour of keeping the global but: 1. Is it used? 2. Do we want it?

docs/entitytypes.wiki states "MediaWiki bindings are needed by defining a content class extending EntityContent and a handler class extending EntityHandler. Furthermore, the handler has to be globally registered in  $wgWBRepoEntityTypes. It is used to create instances of the content class.". I don't know much about content handler but it looks to me that in order to register the handler, 'content-handler-factory-callback' must be provided in the entity type definition, and this does no require a global var (possibly not any more requires it?). So is this point in the docs simply outdated?

Patch-set: 1
Label: Code-Review=0