Update patch set 2
Patch Set 2:

So what concerns me is whether a service like PropertyInfoTable should be constructing the entity ID object from some parts (like a number here), and whether it should be constructing it at all. Wrapping the "constructing from part" in PropertyId::newForeignFromNumber is better as it does not make the service rely on details on how to build the ID but it is still the service that constructs the ID object in the end. This might be OK but it feels a bit wrong to me.
If we want to make sure this class has no way to construct e.g. an item ID, we could then make EntityIdComposer an interface, and a have an implementation like PropertyIdComposer that only deals with proprety IDs injected to PropertyInfoTable. Still I am not saying this is something we should do.
If using a service that know about other entity types that properties is the actual "violation" this patch is intending to fix this could be an alternative solution. I am not sure which is better.

Patch-set: 2
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0