Update patch set 1
Patch Set 1:

Oh, this is indeed bad and shame I missed it. Thanks for spotting.

First regarding some remarks in the commit message:

 > Currently there is only one EntityDataRetrievalServiceFactory implementation. 
 > This means we can as well ditch the interface and type hint against the specific DispatchingServiceFactory implementation.

There is an intention to have a second implementation of EntityDataRetrievalServiceFactory, to be used be WikibaseRepo when not using the factory from Client. It might still take some time before we have it, so I am sorry I cannot provide more than this promise at this point.

As to how to solve this: I also think having an interface that is combining EntityDataRetrievalServiceFactory and the EntityStoreWatcher would be a solution here.

I think adding a check if the factory is an EntityStoreWatcher and skipping updating caches if it is not is something we should NOT to do. We want this "change propagation" here always.

Patch-set: 1
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0