Update patch set 2
Patch Set 2:

> If we want to make sure this class has no way to construct e.g. an item ID, we could then make EntityIdComposer an interface, and a have an implementation like PropertyIdComposer […]

This does not solve the problem I'm concerned about.

Such an interface is of no use in this case. The only guarantee the interface can give is that it returns an EntityId, which causes the type safety violations I'm fixing here. We must use the concrete implementation to "make sure" a class always gets the entity type it expects.

This "concrete implementation" exists. It's called PropertyId::newFromNumber.

Patch-set: 2
Label: Verified=0