Better names for WikibaseRepo/Client (de)serializer factory methods
This patch does not do anything but renaming a few methods in the
WikibaseRepo as well as WikibaseClient top level factories. I'm using
class aliases in the use section as a proof of concept. If these new
class names are fine and this patch got merged, I will rename the
original classes in future releases of the relevant components.

The problem: We do have a "Internal Serialization" and a "DataModel
Serialization" component. The names are confusing.

The so called "DataModel Serialization" component is a "base"
(de)serialization component that only supports Items and Properties.
This is on purpose. We should not teach this component how to deal
with other entity types.

The so called "internal" component really is internal. The only reason
we have it is because there are still old serialization formats in the
Wikidata.org database.

Entity (de)serializers are constructed in the top level factory, which
is fine. The *de*serializer can be passed around as part of the
InternalFormatDeserializerFactory. The *serializer* must be passed around
as a separate $entitySerializer parameter. This is by design and nothing
that needs fixing.

Bug: T160439
Change-Id: Ic126c46d84d15b255a3cac69b3260c08b22fff9f