Update patch set 3
Patch Set 3:

I think it would me much easier that we won't think about how each  service treats OpenGraph, but accept OpenGraph as independent protocol and just implement it (or not).
That way we can save much discussion time.

> Pinterest is a "walled garden" scam I really do not want to support.

That's exactly why it is better to decide to support markup protocol, but not exact service. Many services/sites work correctly with OpenGraph, so why not just stick to the protocol?

Patch-set: 3