Update patch set 1
Patch Set 1:

I know what \v does. This does not mean there is a test missing.

We have dozens of places that use shortcuts for character classes, in most cases it's a \s. It's impossible to have complete test coverage for such code. Asking for something like this just does not make sense. It would mean you have to encode the entirety of the regular expression syntax into the test, if not the entirety of Unicode.

For example, most code that utilizes the \s character class tests it with the space and newline characters as positive, and letters as negative tests. Sure, you could add all characters from the \s character class to the test, as well as all characters that are not in the character class. That would not make anything better, or more robust, or better to read, the contrary.

No. We are testing regular expressions with a few positive and a few negative test cases. This is the case here.

Patch-set: 1