Update patch set 1
Patch Set 1:

> Also this is not a "new constraint". We are disallowing vertical
 > whitespace. That's all.

Vertical whitespace was allowed, now it is not. For me it seems like a new constraint.

> We are testing regular expressions with ...

In this case, I believe we should not test regex, but should test validator (regex is implementation detail in this case). If we state that certain characters are allowed and certain are not, then we should provide either black or white list that describes it.
In case of a black list, I think it should be enforced by tests.
Here, as I understood, is black list case, isn't it?
List isn't probably that big.

PS: Current tests in `ValidatorBuildersTest` pass if I replace regex in `ValidatorBuilders.php:161` with '/\s/m'. Should I just do it then?

Patch-set: 1