Update patch set 8
Patch Set 8:

It's passing now. I wrongly changed a few to return a Promise when in fact their save() call never resolves (the first few tests that verify the API call return an ever-pending deferred from setLabel/setDescription). But I see now that the call in question is always made before any async part so those can stay as-is.

Note that this doesn't convert them all. It's just an example that I'd like you to follow further as this is blocking jQuery 3.0 adoption.

Patch-set: 8