Remove unused Change::getType from the interface
These classes are unpaid technical dept from 2012.

Originally I wanted to introduce a dedicated $action field in the
EntityChange class. This is possible, but would create duplication with
the "type", which also encodes the action.

For now I'm focusing on removing unused code from the interface. The
getType() method is barely used, only for logging. Note the the "type"
field is still there, it is even persisted in the database and can not
be removed!

Change-Id: I40395fd4b3ba65be0cd4063375c5d1086e78e8d6