Update patch set 13
Patch Set 13:

(1 comment)

Looks pretty good!
I've checked that production part of the patch is correctly copied from Api\EditEntity as intended.

I've left some remark on test - I am not sure though if what I suggest there is actually an improvement or not.

Also if I see right the case of editing the statement with the existing id is not tested. This would be a bit more tricky as as actual statement data would need to be compared somehow. But if it is not that much of hassle maybe it would be worth adding to have all three scenarios covered? What is your impression on this?

Patch-set: 13
Label: Verified=0