Remove $message param from ValidatorBuildersTest::provideDataTypeValidation
This is a absolute trivial step in a series of changes I'm doing here. All
this patch does is removing the 4th parameter from all test cases, and moving
it to the front as an array key. I also had to rename a few duplicate array
keys.

This does not make a difference, practically, because these array keys are
nicely printed out when a test fails.

Change-Id: If1c7198ef67ca609d4ab82e72c955641a14b662b