Update patch set 5
Patch Set 5: Code-Review-1

(2 comments)

Alex and I talked about "internal vs. external link syntax". We came to the conclusion that it's much better to stick to the current external link syntax (that is [URL label]). This guarantees this code will always work, even on wikis that do not know anything about Commons.

If this turns out to be a problem (e.g. the community complains that these links are rendered as external links in an other color and with an arrow icon), we can improve this wikitext formatter later to check for the "commons" prefix in the interwiki table, and use it if it's there.

All other previous comments still apply.

Patch-set: 5
Label: Code-Review=-1